src/webapp/requirements.txt
pandas==0.24.2
gevent==1.4.0
numpy==1.16.2
tqdm==4.32.2
Flask==1.1.2
Werkzeug==1.0.1
opencv_python==4.0.0.21
efficientnet==1.1.0
matplotlib==3.2.1
Pillow==7.1.2
scikit_learn==0.22.2.post1
scipy==1.4.1
tensorflow==2.1.0
